Friday and Saturday
Yesterday I fixed most of the animation stuff so all of that works again.
Today I got the video working in the intro scene, so that's great! We can show the video, then fade out and switch to the gamescene. It took me ages trying to fix it with a Ui screen, but finally I gave up and udes a 3d cube instead. then it worked perfectly. flawed system if i have to use 3d in a 2d game but eh, nobody will know